+++ tests/testthat/test_mlr_callbacks.R
@@ -52,7 +52,25 @@ test_that("backup callback works", {
   file = tempfile(fileext = ".rds")
 
   instance = tune(
-    method = "random_search",
+    method = tnr("random_search"),
+    task = tsk("pima"),
+    learner = lrn("classif.rpart", cp = to_tune(1e-04, 1e-1, logscale = TRUE)),
+    resampling = rsmp("cv", folds = 3),
+    measures = msr("classif.ce"),
+    term_evals = 4,
+    batch_size = 2,
+    callbacks = clbk("mlr3tuning.backup", path = file)
+  )
+
+  expect_file_exists(file)
+  expect_benchmark_result(readRDS(file))
+})
+
+test_that("backup callback works with standalone tuner", {
+  file = tempfile(fileext = ".rds")
+
+  instance = tune(
+    method = tnr("grid_search"),
     task = tsk("pima"),
     learner = lrn("classif.rpart", cp = to_tune(1e-04, 1e-1, logscale = TRUE)),
     resampling = rsmp("cv", folds = 3),